Update Code Samples commands (#505)
This commit updates the commands one should use when verifying that
they've properly built the SDK and/or have the SGX properly set up,
up and running.

It turns out that the executable for the samples is
generated inside a `bin` subdir (located in each sample), and it is
imperative to run the sample directly from that dir as otherwise
one will be greeted with `failed to load enclave`-like error.

Signed-off-by: Jakub Konka <jakub.konka@golem.network>