Add the interrupt mechanism for dynamically-loaded workloads
Workloads in an enclave can be classified into two categories:
statically-loaded and dynamically-loaded workloads. Statically-loaded
workloads are application code that are built into the enclave; that is,
they are part of the enclave since enclave initialization.
Dynamically-loaded workloads, as the name suggests, are application code
loaded after the enclave gets running.

One typical example of dynamically-loaded workloads is user programs
loaded by a SGX LibOS. The user programs could be arbitrary code. As a
result, once the user program gets executed, the LibOS may never have
the opportunity to take control of the CPU. Without the ability to
regain the control, it is impossible for the LibOS to implement features
like interruptible signal handler or preemptive in-enclave thread
scheduling.

To address the issue above, we implement the signal-based interrupt
mechanism for dynamically-loaded workloads. With the provided APIs, the
users can now interrupt the dynamically-loaded workloads executed in a
SGX thread by simply sending a real-time POSIX signal (whose number is
64, the max value of signal numbers on Linux) to the SGX thread. The
signal will be captured and (if the timing is good) a pre-registered
interrupt handler will get executed inside the enclave.